refactor(D-168 A1): extract chat_types + chat_context from routes.py (verbatim)
- chat_types.py (leaf): history constants + ChatRunContext + injected-exercise/
  support-level extractors + StreamFrame + MissionEventEnvelope
- chat_context.py: history/digest builders + entity-anchor & ambiguity guard +
  checkpoint history + _canonicalize_mission_event (imports get_checkpointer itself)
- routes.py re-imports every moved name (# noqa: F401 re-exports) so external
  runtime imports and handler-resolved monkeypatches keep working
- lockstep: test_chat_context_seed_strategy patches chat_context.<name>
  (late-binding rule: patch the module where the caller lives)
- manifest: API_SOURCE_FILES += chat_types.py, chat_context.py

routes.py: 3,312 -> 2,749 lines. Gates green, 172 affected tests pass.
